+++ +gams/+transfer/+symbol/Equation.m
@@ -88,6 +88,10 @@ classdef Equation < gams.transfer.symbol.Symbol
         type
     end
 
+    properties (Hidden, Dependent)
+        type_int
+    end
+
     properties (Dependent, SetAccess = private)
         %> Equation default values
 
@@ -103,6 +107,14 @@ classdef Equation < gams.transfer.symbol.Symbol
 
         function obj = set.type(obj, type)
             obj.def_.type = type;
+        end
+
+        function type_int = get.type_int(obj)
+            type_int = obj.def_.type.value;
+        end
+
+        function obj = set.type_int(obj, type_int)
+            obj.def_.type = type_int;
         end
 
         function default_values = get.default_values(obj)
@@ -185,8 +197,8 @@ classdef Equation < gams.transfer.symbol.Symbol
                         index = index + 2;
                         is_pararg = true;
                     elseif strcmpi(varargin{index}, 'records')
-                        obj.data_ = gams.transfer.utils.parse_argument(varargin, ...
-                            index + 1, 'records', @obj.validateData);
+                        records = gams.transfer.utils.parse_argument(varargin, ...
+                            index + 1, 'records', []);
                         has_records = true;
                         index = index + 2;
                         is_pararg = true;
@@ -201,7 +213,9 @@ classdef Equation < gams.transfer.symbol.Symbol
             catch e
                 error(e.message);
             end
-            if ~has_records
+            if has_records
+                obj.setRecords(records);
+            else
                 obj.data_ = gams.transfer.symbol.data.Struct.Empty(obj.def_.domains);
             end
         end