+++ +gams/+transfer/+symbol/Set.m
@@ -170,8 +170,8 @@ classdef Set < gams.transfer.symbol.Symbol
                         index = index + 2;
                         is_pararg = true;
                     elseif strcmpi(varargin{index}, 'records')
-                        obj.data_ = gams.transfer.utils.parse_argument(varargin, ...
-                            index + 1, 'records', @obj.validateData);
+                        records = gams.transfer.utils.parse_argument(varargin, ...
+                            index + 1, 'records', []);
                         has_records = true;
                         index = index + 2;
                         is_pararg = true;
@@ -190,7 +190,9 @@ classdef Set < gams.transfer.symbol.Symbol
             if ~has_domains
                 obj.def_.domains = {gams.transfer.symbol.domain.Relaxed(gams.transfer.Constants.UNIVERSE_NAME)};
             end
-            if ~has_records
+            if has_records
+                obj.setRecords(records);
+            else
                 obj.data_ = gams.transfer.symbol.data.Struct.Empty(obj.def_.domains);
             end
         end